Terminando proyecto

+++ src/main/java/funciones/Average.java
@@ -16,7 +16,7 @@ public class Average extends Function{
     }
 
     @Override
-    public double returnResult() {
+    public Integer returnResult() {
         return 0;
     }
 }
