+++ src/main/java/funciones/Sum.java
@@ -17,7 +17,7 @@ public class Sum extends Function{
     }
 
     @Override
-    public double returnResult() {
+    public Integer returnResult() {
         return 0;
     }
 }
